Gain estimation

Gain estimation

Missing docstring.

Missing docstring for GainEquation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GainEstimationMethod. Check Documenter's build log for details.

FeedbackParticleFilters.solve! — Function.

solve!(eq::GainEquation, method::GainEstimationMethod) --> eq

Solves the gain equation eq using method method.

source

FeedbackParticleFilters.update! — Function.

update!(filter_state, filter_algo, obs, dt) --> filter_state

Updates the filter state by performing one forward step of the model and then assimilating the observation.

source

update!(eq::GainEquation)

Updates the gain equation eq.

update!(eq::GainEquation, ens::ParticleRepresentation)

Updates the gain equation eq by incorporating new information from the ensemble ens.

source

Gain estimation

Poisson equation

Gain estimation methods

Poisson equation

Missing docstring.

Missing docstring for PoissonEquation. Check Documenter's build log for details.

Gain estimation methods

FeedbackParticleFilters.ConstantGainApproximation — Type.

ConstantGainApproximation()

Represents an approximation of the gain by a constant (in the Euclidean sense) vector field, given by the covariance of the observation function and x under the particle distribution.

source

FeedbackParticleFilters.SemigroupMethod — Type.

SemigroupMethod(epsilon, delta, max_iter)

Semigroup method from Algorithm 1 in [1].

[1] Taghvaei, A., & Mehta, P. G. (2016). Gain function approximation in the feedback particle filter. In 2016 IEEE 55th Conference on Decision and Control (CDC) (pp. 5446–5452). IEEE. https://doi.org/10.1109/CDC.2016.7799105

SemigroupMethod(epsilon, delta, max_iter, lambda)

Semigroup method with regularization parameter lambda.

source

FeedbackParticleFilters.DifferentialRKHSMethod — Type.

DifferentialRKHSMethod1d(epsilon, lambda)

Differential loss reproducing kernel Hilbert space (RKHS) method from [1], Section III.

Parameters

epsilon: variance of the Gaussian kernel, Eq. 17 in [1]

lambda: regularization parameter, Eq. 20 in [1]

Reference

[1] Radhakrishnan, A. & and Meyn, S. (2018). Feedback particle filter design using a differential-loss reproducing kernel Hilbert space. 2018 Annual American Control Conference (ACC). IEEE. https://doi.org/10.23919/ACC.2018.8431689

source