Hidden state models

Hidden state models

FeedbackParticleFilters.HiddenStateModel — Type.

HiddenStateModel{S, T<:TimeType} <: AbstractModel{S}

Abstract type for any model of the hidden state of type S.

source

FeedbackParticleFilters.state_dim — Function.

state_dim(problem::AbstractFilteringProblem)

Return the dimensionality of the hidden state in problem.

source

state_dim(model::HiddenStateModel)

Returns the dimensionality of the hidden state in model.

source

FeedbackParticleFilters.state_type — Function.

state_type(arg)

Returns the type of the hidden state in arg. Supported argument types:

Any subtype of AbstractModel

Any subtype of AbstractFilteringProblem

source

state_type(model::HiddenStateModel)

Returns the data type of the hidden state in model.

source

Missing docstring.

Missing docstring for time_type. Check Documenter's build log for details.

FeedbackParticleFilters.initial_condition — Function.

initial_condition(model::HiddenStateModel)

Returns the specification of the initial condition in model. This is either a fixed value or a samplable distribution.

source

FeedbackParticleFilters.initialize — Function.

initialize(model::HiddenStateModel)

Returns a sample from the initial distribution of model.

source

FeedbackParticleFilters.propagate! — Function.

propagate!(state[s], model::HiddenStateModel[, dt])

Propagates the state(s) according to the model. For ContinuousTime' models, a time stepdt' has to be provided. Multiple states are given as a matrix with columns corresponding to states, and are processed i.i.d.

source

propagate!(sfs, filtering_problem, filtering_algorithm; dt) --> sfs

Propagates the system and filter states for one time-step according to the specified filtering problem and algorithm.

source

Hidden state models

Diffusion processes

Diffusion processes

Missing docstring.

Missing docstring for DiffusionStateModel. Check Documenter's build log for details.

FeedbackParticleFilters.LinearDiffusionStateModel — Type.

LinearDiffusionStateModel(A, B; init)

Returns a linear diffusion process hidden state model $dX_t = A X_t dt + B dW_t$ with appropriately sized matrices $A$ and $B$.

Optional argument init stands for the initial condition of the process, which is either

A vector of length n for a fixed (deterministic) initial condition

A Distributions.Sampleable type for a random initial condition

If argument init is left out, it is set to either

a multivariate normal distribution with covariance matrix set to the stationary variance, if it exists

the zero vector

source

Missing docstring.

Missing docstring for ScalarDiffusionStateModel. Check Documenter's build log for details.

FeedbackParticleFilters.drift_function — Function.

drift_function(model)

Returns the drift function $f$ of the diffusion model $dX_t = f(X_t)dt + g(X_t)dW_t$.

source

FeedbackParticleFilters.diffusion_function — Function.

diffusion_function(model)

Returns the diffusion function $g$ of the diffusion model $dX_t = f(X_t)dt + g(X_t)dW_t$.

source

FeedbackParticleFilters.noise_dim — Function.

noise_dim(model)

Returns the dimension of the Brownian motion $W_t$ in the diffusion model $dX_t = f(X_t)dt + g(X_t)dW_t$.

source