Reference

Reference

Contents

Reference

Basic abstractions

Basic methods

Hidden state models

Basic abstractions

FeedbackParticleFilters.AbstractHiddenState — Constant.

AbstractHiddenState = Union{Number, AbstractVector{T}

Abstract type for the state space of the hidden state.

source

FeedbackParticleFilters.VectorHiddenState — Constant.

VectorHiddenState = Union{Float64, AbstractVector{T} where T<:Float64}

Abstract type for those state spaces of the hidden state that shall be treated as vector spaces (i.e. support addition and multiplication by scalars). This excludes points from a manifold, for example.

source

FeedbackParticleFilters.AbstractFilterRepresentation — Type.

AbstractFilterRepresentation{S}

Abstract type for representation of the conditional distribution over the hidden state of type S.

source

FeedbackParticleFilters.ParticleRepresentation — Type.

ParticleRepresentation{S} <: AbstractFilterRepresentation{S}

Abstract type for the representation of the conditional distribution over the hidden state by particles (samples), weighted or unweighted.

source

FeedbackParticleFilters.UnweightedParticleRepresentation — Type.

UnweightedParticleRepresentation{S} <: ParticleRepresentation{S} <: AbstractFilterRepresentation{S}

Abstract type for the representation of the conditional distribution over the hidden state by unweighted particles (samples).

source

FeedbackParticleFilters.AbstractGainEquation — Type.

AbstractGainEquation{T<:AbstractHiddenState}

Abstract type for an equation that determines the gain (vector field).

source

FeedbackParticleFilters.EmptyGainEquation — Type.

EmptyGainEquation{T<:AbstractHiddenState}

An empty gain equation, i.e. no information about the gain is known.

source

Basic methods

Base.eltype — Function.

eltype(filter_rep::T) where T<:AbstractFilterRepresentation{S} where S <: AbstractHiddenState

Extract S.

source

FeedbackParticleFilters.Map — Function.

Map(f::T, A::AbstractArray) where T<:Function

Same as Base.map.

source

Map(F::NTuple{N, Function}, x::T) where {N, T<:Number}

Output an array with the results of applying each f in F to x.

source

Map(F::AbstractArray{Function}, x::T) where {N, T<:Number}

Output an array with the results of applying each f in F to x.

source

Map(F::AbstractArray{U}, x::T) where {N, T<:Number, U<:Function}

Output an array with the results of applying each f in F to x.

source

Map(F::NTuple{N, Function}, A::AbstractArray; output_shape=2) where N

Output an Array of applying each f in F to each a in A, where output_shape determines whether the output should inherit the shape of F or A.

source

Map(F::AbstractArray{T}, A::AbstractArray; output_shape=2) where {N, T<:Function}

Output an Array of applying each f in F to each a in A, where output_shape determines whether the output should inherit the shape of F or A. I.e. if output_shape=1, then for each f in F the output contains an element which contains f applied to all elements from A. If output_shape=2, then for each a in A the output contains an element which contains all f's in F applied to a. Map also implements recursion when output_shape=2: Map starts at the first level and recursively tries to map to each level until a method has been found. If output_shape=1, and error is thrown if the functions in F cannot be applied at the first level of A.

source

Map(f::U, ensemble::T) where {U<:Function, T<:UnweightedParticleRepresentation{S}} where S<:AbstractHiddenState

Short-hand for Base.map applied to ensemble.positions.

source

Map(F::NTuple{N, U}, ensemble::T) where {N, U<:Function, T<:UnweightedParticleRepresentation{S}} where S<:AbstractHiddenState

Short-hand for Map applied to ensemble.positions.

source

Hidden state models

Missing docstring.

Missing docstring for StateModel. Check Documenter's build log for details.

FeedbackParticleFilters.DiffusionStateModel — Type.

DiffusionStateModel(drift_function::Function, diffusion_function::Function, n::Int, nprime::Int) <: ObservationModel

A diffusion process hidden state model $dX_t = f(X_t)dt + g(X_t)dW_t$, where f is the drift_function, g is the observation_function, Xt is the n-dimensional hidden state at time t, and Wt is an nprime-dimensional Brownian motion process.

source

FeedbackParticleFilters.ScalarDiffusionStateModel — Type.

Here's some inline maths: $\sqrt[n]{1 + x + x^2 + \ldots}$.

Here's an equation:

$\frac{n!}{k!(n - k)!} = \binom{n}{k}$

This is the binomial coefficient.

source

FeedbackParticleFilters.Propagator — Function.

Propagator(model::HiddenStateModel)

Returns a function called propagate! that propagates the state model by one time step.

source