Hidden state models

Hidden state models

Missing docstring.

Missing docstring for StateModel. Check Documenter's build log for details.

FeedbackParticleFilters.propagate! — Function.

propagate!(sfs, filtering_problem, filtering_algorithm; dt) --> sfs

Propagates the system and filter states for one time-step according to the specified filtering problem and algorithm.

source

Hidden state models

Diffusion processes

Diffusion processes

Missing docstring.

Missing docstring for DiffusionStateModel. Check Documenter's build log for details.

FeedbackParticleFilters.LinearDiffusionStateModel — Type.

LinearDiffusionStateModel(f::Function, g::Function; init)

A diffusion process hidden state model $dX_t = f(X_t)dt + g(X_t)dW_t$, where f is the drift_function, g is the observation_function, Xt is the n-dimensional hidden state at time t, and Wt is an m-dimensional Brownian motion process.

Optional argument init stands for the initial condition of the process, which is either

A vector of length n for a fixed (deterministic) initial condition

A Distributions.Sampleable type for a random initial condition

If argument init is left out, it is set to either

a multivariate normal distribution with covariance matrix set to the stationary variance, if it exists

the zero vector

source

Missing docstring.

Missing docstring for ScalarDiffusionStateModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drift_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for diffusion_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for initial_condition. Check Documenter's build log for details.