Observation models

Observation models

FeedbackParticleFilters.ObservationModel — Type.

ObservationModel{S1, S2, T} <: AbstractModel{S2}

Abstract type for any model of observations of type S2 and hidden states of type S1.

Calling the model samples a new observation. For discrete-time or static models,

model(x)

returns a new observation based on hidden state x.

For continuous-time models,

model(x, dt)

returns a new observation based on hidden state x and time-step dt.

If x is a vector of state_type elements or a batch matrix, an observation is generated for each element of the batch.

source

Missing docstring.

Missing docstring for emit!. Check Documenter's build log for details.

Observation models

Diffusion processes

Diffusion processes

Missing docstring.

Missing docstring for DiffusionObservationModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LinearDiffusionObservationModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ScalarDiffusionObservationModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for observation_function. Check Documenter's build log for details.