Reference

Reference

Contents

Reference

Basic abstractions

Basic methods

Hidden state models

Basic abstractions

Missing docstring.

Missing docstring for AbstractHiddenState. Check Documenter's build log for details.

Missing docstring.

Missing docstring for VectorHiddenState. Check Documenter's build log for details.

FeedbackParticleFilters.AbstractFilterRepresentation — Type.

AbstractFilterRepresentation{S}

Abstract type for representation of the conditional distribution over the hidden state of type S.

source

FeedbackParticleFilters.ParticleRepresentation — Type.

ParticleRepresentation{S} <: AbstractFilterRepresentation{S}

Abstract type for the representation of the conditional distribution over the hidden state by particles (samples), weighted or unweighted.

source

FeedbackParticleFilters.UnweightedParticleRepresentation — Type.

UnweightedParticleRepresentation{S} <: ParticleRepresentation{S} <: AbstractFilterRepresentation{S}

Abstract type for the representation of the conditional distribution over the hidden state by unweighted (i.e. equally weighted) particles (samples).

source

Missing docstring.

Missing docstring for AbstractGainEquation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EmptyGainEquation. Check Documenter's build log for details.

Basic methods

Base.eltype — Function.

eltype(ensemble::ParticleRepresentation)

Return the type of individual particles in ensemble.

source

Missing docstring.

Missing docstring for Map. Check Documenter's build log for details.

Hidden state models

Missing docstring.

Missing docstring for StateModel. Check Documenter's build log for details.

FeedbackParticleFilters.DiffusionStateModel — Type.

DiffusionStateModel(f::Function, g::Function, init)

A diffusion process hidden state model $dX_t = f(X_t)dt + g(X_t)dW_t$, where f is the drift_function, g is the observation_function, Xt is the n-dimensional hidden state at time t, and Wt is an m-dimensional Brownian motion process.

Argument init stands for the initial condition of the process, which is either

A vector of length n for a fixed (deterministic) initial condition

A Distributions.Sampleable type for a random initial condition

source

Missing docstring.

Missing docstring for ScalarDiffusionStateModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Propagator. Check Documenter's build log for details.